更改了低通滤波器

+++ new-posys/Action_User/develop/temperature_control.h
@@ -8,7 +8,7 @@ int TempErgodic(int gyroNum,int reset);
 void pwm_init(uint32_t arr,uint32_t psc);
 void temp_pid_ctr(int gyro,float val_ex);
 uint32_t Heating(void);
-float LowPassFilter(float newValue,int gyro);
+double LowPassFilter(float newValue,int gyro);
 int HeatingInit(float temp_temp[GYRO_NUMBER]);
 int getTempInitSuces(void);
 void SetTempInitSuces(void);
